NavBar and pages/ setup
