+++ src/pages/PageScales.tsx
@@ -1,0 +1,9 @@
+const PageScales = () => (
+  <div className="p-6">
+    <h1 className="text-3xl font-bold mb-4">Scales</h1>
+    {/*  */}
+  </div>
+)
+
+export default PageScales
+
