+++ src/pages/PageWhatever.tsx
@@ -1,0 +1,9 @@
+const PageWhatever = () => (
+  <div className="p-6">
+    <h1 className="text-3xl font-bold mb-4">Whatever</h1>
+    {/* … */}
+  </div>
+)
+
+export default PageWhatever
+
